pyproject.toml
[project]
name = "weather-health"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "black>=25.9.0",
    "dvc>=3.59.2",
    "evidently>=0.7.6",
    "fastparquet>=2024.11.0",
    "hyperopt>=0.2.7",
    "isort>=7.0.0",
    "jupyterlab>=4.4.3",
    "lightgbm>=4.6.0",
    "matplotlib>=3.10.0",
    "mlflow>=2.22.0",
    "mypy>=1.18.2",
    "numpy>=2.1.3",
    "pandas>=2.2.3",
    "prefect>=3.4.3",
    "prefect-aws>=0.5.10",
    "prettytable>=3.16.0",
    "psycopg>=3.2.9",
    "psycopg-binary>=3.2.9",
    "psycopg2-binary>=2.9.10",
    "pyarrow>=19.0.1",
    "python-dotenv>=1.1.0",
    "ruff>=0.14.1",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "shap>=0.47.2",
    "ydata-profiling>=4.16.1",
]
